Widen search results to a long global doc id
ScoreDoc#doc and FieldDoc become long, and the top-N result path follows:
HitQueue/FieldValueHitQueue tie-breaks, the _doc sort comparator,
TopDocs#merge, IndexSearcher#explain/searchAfter, the rescorers and the
kNN result paths. MaxScoreAccumulator carries a long doc so the cross-leaf
doc-order pruning refinement is preserved. The _doc/FIELD_DOC searchAfter
value is now a Long rather than an Integer.

+++ lucene/core/src/java/org/apache/lucene/document/LatLonPoint.java
@@ -424,7 +424,9 @@ public class LatLonPoint extends Field {
       if (points != null) {
         totalHits += points.getDocCount();
         readers.add(points);
-        docBases.add(leaf.docBase);
+        // NearestNeighbor addresses the composite doc space with int; point-nearest does not
+        // support a reader with more than Integer.MAX_VALUE docs.
+        docBases.add(Math.toIntExact(leaf.docBase));
         liveDocs.add(leaf.reader().getLiveDocs());
       }
     }
